Swap Points and Swap Car next to each drivers name
There's a bit of lag between clicking and the server updating the swapCar and swapPoints variables. Might be abusable if people click fast enough. It updates at the same speed as the joker.jsx numbers do, but since the button is no longer behind a dropdown/modal, you can quickly spam a few swaps and they might all go through.